addroutlayerobject-simplified
